requirements.txt
langchain>=0.1.0
langgraph
langchain-groq
python-dotenv
slack-bolt
pandas
matplotlib
sqlalchemy